RANGER-3809: Dummy impl for RangerKafkaAuthorizer#authorizeByResourceType
Since the current implementation of the acls() call throws
UnsupportedOperationException, it masks an authorization error if a
Kafka client tries to call the InitProducerId API and doesn't have
idempotent_write permission on the cluster nor it has a transactional.id
configured.

Until a proper implementation of the acls() method is done by RANGER-3809
we override authorizeByResourceType to get an access denied on the
client side instead of an exception.

Co-authored-by: Viktor Somogyi-Vass <viktorsomogyi@gmail.com>

Change-Id: I20f498bb39edf5f6cc5897fbb6b3d6435bf0c6b5
Signed-off-by: Ramesh Mani <rmani@cloudera.com>
(cherry picked from commit 0580e513e36196a65bd68562b18ef4dddf83ee37)